+++ flexmeasures/cli/tests/conftest.py
@@ -114,20 +114,34 @@ def reporter_config_raw(app, db, setup_dummy_data):
     return reporter_config_raw
 
 
-@pytest.fixture(scope="module")
 @pytest.mark.skip_github
-def setup_dummy_asset(db, app):
+@pytest.fixture(scope="module")
+def process_power_sensor(db, app, add_market_prices):
     """
-    Create an Asset to add sensors to and return the id.
+    Create an asset of type "process", power sensor to hold the result of
+    the scheduler and price data consisting of 8 expensive hours, 8 cheap hours, and again 8 expensive hours-
+
     """
-    dummy_asset_type = GenericAssetType(name="DummyGenericAssetType")
 
-    db.session.add(dummy_asset_type)
+    process_asset_type = GenericAssetType(name="process")
 
-    dummy_asset = GenericAsset(
-        name="DummyGenericAsset", generic_asset_type=dummy_asset_type
+    db.session.add(process_asset_type)
+
+    process_asset = GenericAsset(
+        name="Test Process Asset", generic_asset_type=process_asset_type
     )
-    db.session.add(dummy_asset)
+
+    db.session.add(process_asset)
+
+    power_sensor = Sensor(
+        "power",
+        generic_asset=process_asset,
+        event_resolution=timedelta(hours=1),
+        unit="MW",
+    )
+
+    db.session.add(power_sensor)
+
     db.session.commit()
 
-    return dummy_asset.id
+    yield power_sensor.id